code monitoring: prep for experimental getting started page
