+++ client/web/src/enterprise/code-monitoring/CodeMonitoringPage.tsx
@@ -1,7 +1,5 @@
-import classnames from 'classnames'
-import * as H from 'history'
 import PlusIcon from 'mdi-react/PlusIcon'
-import React, { useCallback, useMemo, useState, useEffect } from 'react'
+import React, { useMemo, useEffect } from 'react'
 import { catchError, map, startWith } from 'rxjs/operators'
 
 import { LoadingSpinner } from '@sourcegraph/react-loading-spinner'
@@ -9,13 +7,11 @@ import { Link } from '@sourcegraph/shared/src/components/Link'
 import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { asError, isErrorLike } from '@sourcegraph/shared/src/util/errors'
 import { useObservable } from '@sourcegraph/shared/src/util/useObservable'
-import { Container, PageHeader } from '@sourcegraph/wildcard'
+import { PageHeader } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../../auth'
 import { CodeMonitoringLogo } from '../../code-monitoring/CodeMonitoringLogo'
-import { FilteredConnection } from '../../components/FilteredConnection'
 import { PageTitle } from '../../components/PageTitle'
-import { CodeMonitorFields, ListUserCodeMonitorsResult, ListUserCodeMonitorsVariables } from '../../graphql-operations'
 import { Settings } from '../../schema/settings.schema'
 import { eventLogger } from '../../tracking/eventLogger'
 
@@ -23,38 +19,22 @@ import {
     fetchUserCodeMonitors as _fetchUserCodeMonitors,
     toggleCodeMonitorEnabled as _toggleCodeMonitorEnabled,
 } from './backend'
-import { CodeMonitorNode, CodeMonitorNodeProps } from './CodeMonitoringNode'
+import { CodeMonitoringGettingStarted } from './CodeMonitoringGettingStarted'
+import { CodeMonitorList } from './CodeMonitorList'
 
 export interface CodeMonitoringPageProps extends SettingsCascadeProps<Settings> {
     authenticatedUser: AuthenticatedUser
-    location: H.Location
-    history: H.History
-
     fetchUserCodeMonitors?: typeof _fetchUserCodeMonitors
     toggleCodeMonitorEnabled?: typeof _toggleCodeMonitorEnabled
 }
 
-type CodeMonitorFilter = 'all' | 'user'
-
 export const CodeMonitoringPage: React.FunctionComponent<CodeMonitoringPageProps> = ({
-    history,
-    location,
     settingsCascade,
     authenticatedUser,
     fetchUserCodeMonitors = _fetchUserCodeMonitors,
     toggleCodeMonitorEnabled = _toggleCodeMonitorEnabled,
 }) => {
     useEffect(() => eventLogger.logViewEvent('CodeMonitoringPage'), [])
-
-    const queryConnection = useCallback(
-        (args: Partial<ListUserCodeMonitorsVariables>) =>
-            fetchUserCodeMonitors({
-                id: authenticatedUser.id,
-                first: args.first ?? null,
-                after: args.after ?? null,
-            }),
-        [authenticatedUser, fetchUserCodeMonitors]
-    )
 
     const LOADING = 'loading' as const
 
@@ -73,16 +53,6 @@ export const CodeMonitoringPage: React.FunctionComponent<CodeMonitoringPageProps
             [authenticatedUser.id, fetchUserCodeMonitors]
         )
     )
-
-    const [monitorListFilter, setMonitorListFilter] = useState<CodeMonitorFilter>('all')
-
-    const setAllFilter = useCallback<React.MouseEventHandler>(() => {
-        setMonitorListFilter('all')
-    }, [])
-
-    const setUserFilter = useCallback<React.MouseEventHandler>(() => {
-        setMonitorListFilter('user')
-    }, [])
 
     return (
         <div className="code-monitoring-page">
@@ -117,107 +87,9 @@ export const CodeMonitoringPage: React.FunctionComponent<CodeMonitoringPageProps
                 className="mb-3"
             />
             {userHasCodeMonitors === 'loading' && <LoadingSpinner />}
-            {!userHasCodeMonitors && (
-                <div className="mt-5">
-                    <div className="d-flex flex-column mb-5">
-                        <h2>Get started with code monitoring</h2>
-                        <p className="text-muted code-monitoring-page__start-subheading mb-4">
-                            Watch your code for changes and trigger actions to get notifications, send webhooks, and
-                            more. <a href="https://docs.sourcegraph.com/code_monitoring">Learn more.</a>
-                        </p>
-                        <Link to="/code-monitoring/new" className="code-monitoring-page__start-button btn btn-primary">
-                            Create your first code monitor →
-                        </Link>
-                    </div>
-                    <div className="code-monitoring-page__start-points container">
-                        <h3 className="mb-3">Starting points for your first monitor</h3>
-                        <div className="row no-gutters code-monitoring-page__start-points-panel-container mb-3">
-                            <div className="col-6">
-                                <div className="card">
-                                    <div className="card-body p-3">
-                                        <h3>Watch for AWS secrets in commits</h3>
-                                        <p className="text-muted">
-                                            Use a search query to watch for new search results, and choose how to
-                                            receive notifications in response.
-                                        </p>
-                                        <a
-                                            href="https://docs.sourcegraph.com/code_monitoring/how-tos/starting_points#watch-for-potential-secrets"
-                                            className="btn btn-secondary"
-                                        >
-                                            View in docs →
-                                        </a>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className="col-6">
-                                <div className="card">
-                                    <div className="card-body p-3">
-                                        <h3>Watch for new consumers of deprecated methods</h3>
-                                        <p className="text-muted">
-                                            Keep an eye on commits with new consumers of deprecated methods to keep your
-                                            code base up-to-date.
-                                        </p>
-                                        <a
-                                            href="https://docs.sourcegraph.com/code_monitoring/how-tos/starting_points#watch-for-consumers-of-deprecated-endpoints"
-                                            className="btn btn-secondary"
-                                        >
-                                            View in docs →
-                                        </a>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                        <a className="link" href="https://docs.sourcegraph.com/code_monitoring/how-tos/starting_points">
-                            Find more starting points in the docs
-                        </a>
-                    </div>
-                    <div className="code-monitoring-page__learn-more container mt-5">
-                        <h3 className="mb-3">Learn more about code monitoring</h3>
-                        <div className="row">
-                            <div className="col-4">
-                                <div>
-                                    <h4>Core concepts</h4>
-                                    <p className="text-muted">
-                                        Craft searches that will monitor your code and trigger actions.{' '}
-                                        <a
-                                            href="https://docs.sourcegraph.com/code_monitoring/explanations/core_concepts"
-                                            className="link"
-                                        >
-                                            Read the docs
-                                        </a>
-                                    </p>
-                                </div>
-                            </div>
-                            <div className="col-4">
-                                <div>
-                                    <h4>Starting points and ideas</h4>
-                                    <p className="text-muted">
-                                        Find specific examples of useful code monitors to keep on top of security and
-                                        consistency concerns.{' '}
-                                        <a
-                                            href="https://docs.sourcegraph.com/code_monitoring/how-tos/starting_points"
-                                            className="link"
-                                        >
-                                            Explore starting points
-                                        </a>
-                                    </p>
-                                </div>
-                            </div>
-                            <div className="col-4">
-                                <div>
-                                    <h4>Questions and feedback</h4>
-                                    <p className="text-muted">
-                                        We want to hear your feedback.{' '}
-                                        <a href="mailto:feedback@sourcegraph.com" className="link">
-                                            Share your thoughts
-                                        </a>
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            )}
+
+            {!userHasCodeMonitors && <CodeMonitoringGettingStarted />}
+
             {userHasCodeMonitors && userHasCodeMonitors !== 'loading' && !isErrorLike(userHasCodeMonitors) && (
                 <>
                     <div className="d-flex flex-column">
@@ -232,67 +104,12 @@ export const CodeMonitoringPage: React.FunctionComponent<CodeMonitoringPageProps
                                 </div>
                             </div>
                         </div>
-                        <div className="row mb-5">
-                            <div className="d-flex flex-column col-2 mr-2">
-                                <h3>Filters</h3>
-                                <button
-                                    type="button"
-                                    className={classnames('btn text-left', {
-                                        'btn-primary': monitorListFilter === 'all',
-                                    })}
-                                    onClick={setAllFilter}
-                                >
-                                    All
-                                </button>
-                                <button
-                                    type="button"
-                                    className={classnames('btn text-left', {
-                                        'btn-primary': monitorListFilter === 'user',
-                                    })}
-                                    onClick={setUserFilter}
-                                >
-                                    Your code monitors
-                                </button>
-                            </div>
-                            <div className="d-flex flex-column w-100 col">
-                                <h3 className="mb-2">
-                                    {`${monitorListFilter === 'all' ? 'All code monitors' : 'Your code monitors'}`}
-                                </h3>
-                                <Container>
-                                    <FilteredConnection<
-                                        CodeMonitorFields,
-                                        Omit<CodeMonitorNodeProps, 'node'>,
-                                        (ListUserCodeMonitorsResult['node'] & { __typename: 'User' })['monitors']
-                                    >
-                                        location={location}
-                                        history={history}
-                                        defaultFirst={10}
-                                        queryConnection={queryConnection}
-                                        hideSearch={true}
-                                        nodeComponent={CodeMonitorNode}
-                                        nodeComponentProps={{
-                                            authenticatedUser,
-                                            location,
-                                            showCodeMonitoringTestEmailButton:
-                                                (!isErrorLike(settingsCascade.final) &&
-                                                    settingsCascade.final?.experimentalFeatures
-                                                        ?.showCodeMonitoringTestEmailButton) ||
-                                                false,
-                                            toggleCodeMonitorEnabled,
-                                        }}
-                                        noun="code monitor"
-                                        pluralNoun="code monitors"
-                                        noSummaryIfAllNodesVisible={true}
-                                        cursorPaging={true}
-                                        className="filtered-connection__centered-summary"
-                                    />
-                                </Container>
-                            </div>
-                        </div>
-                        <div className="mt-5">
-                            We want to hear your feedback!{' '}
-                            <a href="mailto:feedback@sourcegraph.com">Share your thoughts</a>
-                        </div>
+                        <CodeMonitorList
+                            settingsCascade={settingsCascade}
+                            authenticatedUser={authenticatedUser}
+                            fetchUserCodeMonitors={fetchUserCodeMonitors}
+                            toggleCodeMonitorEnabled={toggleCodeMonitorEnabled}
+                        />
                     </div>
                 </>
             )}
